♻️  Refactor Validator
Validator class의 이름을 변경하고 do-catch에러를 try? 로 변경했다.
