+++ SignUp/SignUp/Delegate/TextFiledDelegate/PwTextFieldDelegate.swift
@@ -14,7 +14,7 @@ class PwTextFieldDelegate: BaseTextFieldDelegate {
     func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange, replacementString string: String) -> Bool {
         let text = (textField.text ?? "") as NSString
         let newText = text.replacingCharacters(in: range, with: string)
-        let isCorrect = Validater.isCorrected(pw: newText)
+        let isCorrect = Validator.isCorrected(pw: newText)
         let result = isCorrect
 
             isCorrect.keys.forEach{ isCorrect in
